+++ src/laplacian.cpp
@@ -98,3 +98,30 @@ Rcpp::NumericMatrix outer_cpp(const std::vector<double>& xx,
   }
   return out;
 }
+
+
+
+// [[Rcpp::export]]
+Rcpp::NumericMatrix prep_lapl_spec(const Rcpp::List& phy) {
+  auto edge = phy_to_edge(phy);
+  auto el   = phy_to_el(phy);
+
+  auto num_nodes = phy["Nnode"];
+  Rcpp::StringVector tips = phy["tip.label"];
+  auto num_tips = tips.size();
+
+  std::vector< std::vector< double >> lapl_mat = dist_nodes(edge,
+                                                            el,
+                                                            num_tips,
+                                                            num_nodes);
+  Rcpp::NumericMatrix res(lapl_mat.size(), lapl_mat[0].size());
+
+  for (size_t i = 0; i < lapl_mat.size(); ++i) {
+    for (size_t j = 0; j < lapl_mat[i].size(); ++j) {
+      res(i, j) = lapl_mat[i][j];
+    }
+    res(i, i) = - std::accumulate(lapl_mat[i].begin(), lapl_mat[i].end(), 0.0);
+  }
+
+  return res;
+}