+++ src/phylo.cpp
@@ -23,6 +23,7 @@
 #include "L2newick.h"    // NOLINT [build/include_subdir]
 
 #include "imbalance_steps.h"  // NOLINT [build/include_subdir]
+#include "branch_colless.h"  // NOLINT [build/include_subdir]
 
 //' Function to generate an ltable from a phy object.
 //' @description This function is a C++
@@ -81,7 +82,15 @@ double imbalance_steps_cpp(const Rcpp::NumericMatrix& ltable_R,
     } catch (const char* msg) {
        Rcpp::Rcout << msg << std::endl;
     } catch(...) {
-       ::Rf_error("c++ exception (unknown reason)");
+      Rcpp::stop("c++ exception (unknown reason)");
     }
     return NA_REAL;
 }
+
+// [[Rcpp::export]]
+double calc_branch_colless_cpp(const std::vector<int>& edge,
+                          const std::vector<double>& el) {
+  branch_colless::phylo_tree focal_tree(edge, el);
+  return focal_tree.calculate_branch_colless();
+}
+